requirements.txt
setuptools==44.0.0
Django==1.10
PyYAML==5.1
argparse==1.2.1
asn1crypto==0.24.0
astroid==1.6.5
backports.functools-lru-cache==1.5
configparser==3.5.0
cryptography==2.2.2
django-cors-headers==2.4.0
django-rest-swagger==0.3.5
django-templated-mail==1.1.1
djangorestframework==3.8.2
djoser==1.1.5
enum==0.4.6
enum34==1.1.6
futures==3.2.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.5
pyOpenSSL==18.0.0
pylint==1.9.3
singledispatch==3.4.0.3
six==1.12.0
wrapt==1.10.11
wsgiref==0.1.2
